Fixes, still need to do tests
