+++ src/shared/featureToggle.ts
@@ -3,14 +3,17 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import { SettingsConfiguration } from './settingsConfiguration'
+import { extensionSettingsPrefix } from './constants'
+import { SettingsConfiguration, DefaultSettingsConfiguration } from './settingsConfiguration'
 
-// This enum represents the list of active feature flags to check for
-// Features all need to be added with matching enums and string values
-// e.g. `Feature1 = 'Feature1'`
-// Features that do not follow this scheme will not work.
-// You cannot have more active features than FeatureToggle.maxFeatures (default: 5)
-// Any features that are flagged in the code but not added here will always return false.
+/**
+ * This enum represents the list of active feature flags to check for
+ * Features all need to be added with matching enums and string values
+ * e.g. `Feature1 = 'Feature1'`
+ * Features that do not follow this scheme will not work.
+ * You cannot have more active features than FeatureToggle.maxFeatures (default: 5)
+ * Any features that are flagged in the code but not added here will always return false.
+ */
 export enum ActiveFeatureKeys {
     CloudWatchLogs = 'CloudWatchLogs',
 }
@@ -48,7 +51,10 @@ export class FeatureToggle {
 
     private static INSTANCE: FeatureToggle | undefined
 
-    public constructor(configuration: SettingsConfiguration, overrideKeys?: string[]) {
+    public constructor(
+        configuration: SettingsConfiguration = new DefaultSettingsConfiguration(extensionSettingsPrefix),
+        overrideKeys?: string[]
+    ) {
         this.enabledFeatures = new Set()
 
         let keys: string[] = []
@@ -85,18 +91,9 @@ export class FeatureToggle {
         return this.enabledFeatures.has(key)
     }
 
-    public static createFeatureToggle(configuration: SettingsConfiguration, overrideKeys?: string[]): FeatureToggle {
-        if (FeatureToggle.INSTANCE) {
-            throw new Error('FeatureToggle already initialized')
-        }
-
-        FeatureToggle.INSTANCE = new FeatureToggle(configuration, overrideKeys)
-        return FeatureToggle.INSTANCE
-    }
-
     public static getFeatureToggle(): FeatureToggle {
         if (!FeatureToggle.INSTANCE) {
-            throw new Error('FeatureToggle not initialized')
+            FeatureToggle.INSTANCE = new FeatureToggle()
         }
 
         return FeatureToggle.INSTANCE
